+++ domain/src/main/java/com/sample/domain/model/Rating.kt
@@ -2,5 +2,5 @@ package com.sample.domain.model
 
 data class Rating(
     val count: Int,
-    val rate: Double
+    val rate: Float
 )
